# (performance): Replaced CodeSandbox embed with Sandpack integration