--- a/package.json
+++ b/package.json
@@ -27,6 +27,7 @@
     "@chakra-ui/skip-nav": "^1.2.1",
     "@chakra-ui/theme-tools": "1.3.1",
     "@chakra-ui/utils": "1.9.1",
+    "@codesandbox/sandpack-react": "^0.10.6",
     "@docsearch/react": "^1.0.0-alpha.27",
     "@docusaurus/utils": "^2.0.0-alpha.69",
     "@emotion/react": "^11.7.0",