--- a/src/components/sandpack-embed/index.tsx
+++ b/src/components/sandpack-embed/index.tsx
@@ -1,14 +1,17 @@
-import { useColorMode, Box, BoxProps } from '@chakra-ui/react'
+import { Box, BoxProps } from '@chakra-ui/react'
 import { Sandpack, SandpackProps } from '@codesandbox/sandpack-react'
 import '@codesandbox/sandpack-react/dist/index.css'
 
 const SandpackEmbed = (props: BoxProps & SandpackProps) => {
-  const { colorMode } = useColorMode()
   return (
     <Box
       as={Sandpack}
       {...props}
-      theme={colorMode === 'light' ? 'github-light' : 'night-owl'}
+      options={{
+        ...props.options,
+        showLineNumbers: true,
+      }}
+      theme='dark'
       template='react-ts'
       customSetup={{
         dependencies: {