… configs/sandpack-contents/homepage/files.ts
… yarn.lock