## 📝 Description

I've noticed in PageSpeed Insights that a big part of the network payload comes from the CodeSandbox embed on the homepage. I'm replacing the `iframe` with a Sandpack integration to avoid most of the payload.

## ⛳️ Current behavior (updates)

<img width="960" alt="CleanShot 2021-12-18 at 11 00 10@2x" src="https://user-images.githubusercontent.com/5396211/146637094-1b0d523f-19c8-411d-86b2-d15caa801772.png">

<img width="960" alt="CleanShot 2021-12-18 at 11 05 00@2x" src="https://user-images.githubusercontent.com/5396211/146637204-1376c9c3-e1ec-4a4c-b069-07f8379e2da3.png">


## 🚀 New behavior

All of the metrics have drastically been improved:
- FCP: 0.6s -> 0.4s
- Speed Index: 1.7s -> 0.6s ⭐️⭐️
- LCP: 0.6s -> 0.5s
- TTI: 9.8s -> 1.6s ⭐️⭐️⭐️
- Total Blocking Time: 4.9s -> 0.3s ⭐️⭐️⭐️
- Cumulative Layout Shift: 0.029 -> 0 ⭐️
- Overall Score: 58 -> 86 ⭐️⭐️⭐️

<img width="960" alt="CleanShot 2021-12-18 at 11 05 41@2x" src="https://user-images.githubusercontent.com/5396211/146637219-062e0d6b-6c84-4f51-af87-5d77486e50b0.png">

## 💣 Is this a breaking change (Yes/No):

No